requirements.txt
python-telegram-bot==20.7
   websockets==12.0